requirements.txt
fastapi~=0.110.0
uvicorn[standard]~=0.29.0
slowapi~=0.1.9
python-dotenv~=1.0.1
pydantic~=2.6.4
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4